End of times
Also whoever is stupid enough to actually believe  5000mhz radio waves are causing this.? It’s a wonder you’ve made it this far in life hahaha. What they should have done is create a virus that only targets the stupid people on this world, the only reason dumb shits like that are still around is because we’re nice enough to put warning labels on hazardous product. Otherwise poor little Timmy might have accidentally mistook the bleach for water. Sure good thing we put warning labels on it because Timmy wouldn’t have been smart enough to tell the difference. 😆